python/requirements.txt
opentelemetry-api>=1.21.0
opentelemetry-sdk>=1.21.0
opentelemetry-exporter-otlp-proto-http>=1.21.0
opentelemetry-instrumentation>=0.42b0
